requirements.txt
Pygments==2.10.0
Rx==1.6.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
caio==0.9.3
certifi==2021.10.8
charset-normalizer==2.0.9
colorful==0.5.4
colorlog==6.6.0
frozenlist==1.2.0
graphql-core==3.1.7
idna==3.3
multidict==5.2.0
promise==2.3
pure25519==0.0.1
requests==2.26.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2

absl-py==1.0.0
aiofile==3.7.2
gql==3.0.0rc0  # Be careful of accidental regressions to 2.x
hikari==2.0.0.dev104
lxml==4.7.1  # Hidden dep
prettyprinter==0.18.0
pytz==2021.3
ruamel.yaml==0.17.17
uvloop==0.16.0; sys_platform != "win32"
